MacOSX10.6.sdk

+++ System/Library/Frameworks/IOBluetooth.framework/Versions/A/Headers/objc/IOBluetoothSDPUUID.h
@@ -132,7 +132,7 @@
     @result Returns the length (in bytes) of the UUID.
 */
 
-- (unsigned)length;
+- (NSUInteger)length;
 
 /*!
     @method getUUIDWithLength:
